sender/requirements.txt
Flask==3.0.1
requests==2.26.0
python-dotenv